fix the landing page and directories
