--- src/app/home/components/Courses.module.css
@@ -1,39 +1,0 @@
-.container {
-    position: relative;
-    border: 2px solid;
-    width: 20rem;
-    height: 16rem;
-    margin: 1.25rem;
-    padding: 1.25rem;
-    border-radius: 0.75rem;
-}
-
-/* Add a pseudo-element to handle the hover state */
-.container::after {
-    content: '';
-    position: absolute;
-    top: 2px; /* border width */
-    left: 2px; /* border width */
-    right: 2px; /* border width */
-    bottom: 2px; /* border width */
-    cursor: pointer;
-    border-radius: 0.5rem; /* slightly smaller to fit inside border */
-}
-
-.title {
-    line-height: 2rem;
-    font-weight: bold;
-}
-
-.progressWrapper {
-    margin-top: 0.5rem;
-}
-
-.absoluteBox {
-    position: absolute;
-    width: 8rem;
-    height: 7rem;
-    border: 2px solid;
-    bottom: 1.25rem;
-    right: 1.25rem;
-}
