--- src/app/home/components/DailyQuest.module.css
@@ -1,18 +1,0 @@
-.container {
-    position: absolute;
-    right: 0;
-    border: 2px solid;
-    border-radius: 0.75rem;
-    padding: 1.25rem;
-    width: 25rem; /* Doubled from 20rem to 40rem */
-    height: 40rem;
-    margin: 1.25rem;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 1rem;
-    overflow-y: auto;
-}
-.container:hover {
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.15);
-}
