--- src/app/home/components/LessonBanner.module.css
@@ -1,10 +1,0 @@
-.bannerBox {
-    position: fixed;
-    top: 20px;
-    border: 2px solid black;
-    border-radius: 5px;
-    width: 300px;
-    text-align: left;
-    padding: 10px;
-    left: 50%;
-}
