--- src/app/home/components/LessonCircle.module.css
@@ -1,15 +1,0 @@
-.container {
-    position: relative;
-    left: 50%;
-    top: 200px;
-    border: 1px solid black;
-    cursor: 'pointer';
-    width: 100px;
-    height: 100px;
-    text-align: center;
-    border-radius: 50%;
-    justify-content: center;
-    padding: 20px 1px;
-    overflow: hidden;
-    font-size: 15px;
-}
