--- src/app/home/components/LessonTile.module.css
@@ -1,34 +1,0 @@
-.container {
-    position: relative;
-    cursor: pointer;
-    border-bottom: 1px solid #29354e;
-    background-color: #0c1321;
-    width: 280px;
-    height: 80px;
-    display: 'flex';
-    flex-direction: 'row';
-    align-items: 'center';
-    padding: 1.25rem;
-    display: 'flex';
-    flex-direction: 'row';
-    align-items: 'center';
-    color: white;
-}
-
-.selectedContainer {
-    background-color: #f0f0f0;
-    color: black;
-}
-
-.title {
-    line-height: 2rem;
-    /* equivalent to leading-8 */
-    font-weight: 500;
-    /* equivalent to font-bold */
-    font-size: 20px;
-}
-
-.subtitle {
-    line-height: 1rem;
-    font-size: 12px;
-}
