--- src/app/home/components/LevelIcon.module.css
@@ -1,29 +1,0 @@
-@keyframes pulse {
-    0% {
-      transform: scale(1);
-      opacity: 1;
-    }
-    50% {
-      transform: scale(1.15);
-      opacity: 0.75;
-    }
-    100% {
-      transform: scale(1);
-      opacity: 1;
-    }
-  }
-  
-  .levelIcon {
-    transition: transform 0.3s ease;
-    transform-origin: center center;
-  }
-  
-  .levelIcon:hover {
-    transform: scale(1.2);
-    cursor: pointer;
-  }
-  
-  .pulse {
-    animation: pulse 1.5s infinite;
-  }
-
