--- src/app/home/components/SimplePage.module.css
@@ -1,79 +1,0 @@
-/* @import url('../node_modules/@fontsource/roboto-condensed/index.css');
-@import url('https://fonts.googleapis.com/css2?family=Roboto+Condensed:wght@700&display=swap'); */
-
-.title {
-    font-size: 30px;
-    font-weight: 700;
-}
-.about {
-    font-size: 24px;
-    font-weight: 500;
-}
-.hrContainer {
-    position: relative;
-    height: 5px;
-    margin-top: 5px;
-}
-
-.blackLine {
-    position: absolute;
-    top: 0;
-    left: 0;
-    height: 5px; /* No quotes */
-    width: 54px; /* Width of the black line */
-    background-color: black;
-    border: none;
-    z-index: 2;
-    border-radius: 7px; /* Shorthand for all corners */
-}
-
-.greyLine {
-    position: absolute;
-    top: 50%;
-    left: 0;
-    width: 100%;
-    height: 3px;
-    background-color: #c0c0c0;
-    border: none;
-    z-index: 1;
-    transform: translateY(-50%);
-    border-radius: 7px; /* Shorthand for all corners */
-}
-
-.indent {
-    margin-left: 40px;
-}
-
-.buttonContainer {
-    display: flex;
-    justify-content: flex-end;
-    width: 100%;
-}
-
-.button {
-    border: 1px solid #a0a0a0;
-    height: '50px';
-    margin: 2rem 1.25rem;
-    padding: 11px 32px 11px 33px;
-    border-radius: 0.75rem;
-    background-color: #f0f0f0;
-}
-
-.completeButton {
-    margin: 2rem 1.25rem;
-    padding: 0.5rem;
-    border-radius: 25px;
-    background-color: #f0f0f0;
-}
-
-.activeButton {
-    background-color: #00c7ad; /* Darker color */
-    color: white;
-    cursor: pointer;
-}
-
-.button:disabled {
-    background-color: rgb(220, 220, 220);
-    cursor: not-allowed;
-    opacity: 0.5;
-}
