--- src/app/home/components/StoreItem.module.css
@@ -1,4 +1,0 @@
-.storeItem {
-    border: 1px solid black;
-    padding: 2px;
-}
