--- src/app/home/opioid/lessons.ts
@@ -1,398 +1,0 @@
-const lessons = [
-    {
-        title: 'What Are Opioids?',
-        content: [
-            {
-                text: 'Opioids bind to opioid receptors in the brain resulting in pain relief and pleasurable effects. When bound, opioids also control breathing and feelings of pain and stress.',
-            },
-            {
-                text: 'Examples of opioids include codeine, morphine, oxycontin, fentanyl, Percocet, heroin, Vicodin, and methadone.',
-            },
-            {
-                text: 'Over 100,000 deaths every year in the United States are related to overdoses, and the majority of overdoses result from synthetic opioids like fentanyl (CDC)',
-            },
-        ],
-    },
-    {
-        title: 'What Is Fentanyl?',
-        content: [
-            {
-                text: 'Fentanyl is an opioid painkiller that can be used in medical settings for surgery or treating severe pain.',
-                subpoints: [],
-            },
-            {
-                text: 'Fentanyl has become increasingly found contaminating other substances manufactured and purchased outside of regulated settings.',
-                subpoints: [
-                    {
-                        text: 'The fentanyl often found mixed into other drugs is illegally manufactured, meaning there is no regulation of its strength which can make it extremely addictive and lethal.',
-                        subpoints: [],
-                    },
-                    {
-                        text: 'Fentanyl is approximately 50-100 times stronger than morphine (NIDA).',
-                        subpoints: [],
-                    },
-                ],
-            },
-            {
-                text: 'Fentanyl analogs are the different variations of fentanyl. They often have similar but different chemical structures and may be produced in illegal settings to increase strength compared to medical-grade fentanyl.',
-                subpoints: [
-                    {
-                        text: 'On average it takes approximately 2mg of fentanyl for a person to overdose, but exact amount can depend on the analog of fentanyl ingested, the person’s tolerance to fentanyl/opioids, and the method of ingestion.',
-                        subpoints: [],
-                    },
-                    {
-                        text: 'For an individual who has no tolerance to opioids, a small dose of fentanyl can be extremely lethal.',
-                        subpoints: [],
-                    },
-                ],
-            },
-        ],
-    },
-    {
-        title: 'What Is Opioid Overdose?',
-        content: [
-            {
-                text: 'When too many opioids are ingested or a combination of opioids and other substances are ingested, it can cause an overload of activity from opioid receptors in the brain. Due to the control opioids have over breathing, this can cause a person to enter respiratory depression.',
-                subpoints: [
-                    {
-                        text: 'Respiratory depression is when breathing is too slow or shallow for the exchange of carbon dioxide and oxygen to occur, and it is the life-threatening symptom of opioid overdoses. The buildup of carbon dioxide and lack of oxygen can cause brain damage within 4 minutes, and this is soon followed by death.',
-                        subpoints: [],
-                    },
-                ],
-            },
-        ],
-    },
-    {
-        title: 'What Is Naloxone?',
-        content: [
-            {
-                video: {
-                    pageTitle: 'What Is Naloxone?',
-                    videoPath:
-                        'https://youtu.be/dPS-F91Mkg0?si=HWYS1tzyzssmqhwo',
-                    startTime: '00:00',
-                    endTime: '05:00',
-                },
-            },
-            {
-                text: 'Naloxone is a medication used to reverse an opioid overdose by restoring a person’s breathing within 2-3 minutes of administration.',
-                subpoints: [
-                    {
-                        text: 'Multiple doses may be needed if breathing does not restore 2-3 minutes after administering a dose.',
-                        subpoints: [],
-                    },
-                ],
-            },
-            {
-                text: 'Naloxone is the generic name for the medication that reverses opioid overdoses. Naloxone may come in the form of an intranasal spray or intramuscular injection using a medication vial and syringe. The intranasal spray version of naloxone is made by many different companies and is sometimes referred to by brand names like “Narcan”.',
-                subpoints: [
-                    {
-                        text: 'Different forms and brands of naloxone all accomplish the same goal of reversing the effects of an opioid overdose.',
-                        subpoints: [],
-                    },
-                ],
-            },
-            {
-                text: 'Naloxone’s effects are only temporary. Naloxone can wear off 30-90 minutes after administration, and a person may fall back into an overdose.',
-                subpoints: [
-                    {
-                        text: 'This is why it is always important to call 911 when you see someone overdosing, even if you plan to give them naloxone.',
-                        subpoints: [],
-                    },
-                ],
-            },
-        ],
-    },
-    {
-        title: 'Responding to an Opioid Overdose',
-        content: [
-            {
-                video: {
-                    pageTitle: 'Opioid overdose triad',
-                    videoPath: 'https://www.youtube.com/watch?v=pS5UVJ19VX0',
-                    startTime: '00:00',
-                    endTime: '0:59',
-                },
-            },
-            {
-                text: 'The first step in responding to an opioid overdose is recognizing the signs and symptoms.',
-                subpoints: [
-                    {
-                        text: 'The three main signs of an opioid overdose or the “opioid overdose triad” are:',
-                        subpoints: [
-                            {
-                                text: '**Pinpoint or very small, constricted pupils.** The pupils will not respond to light and will remain small whether a light is on or off of them.',
-                                subpoints: [],
-                            },
-                            {
-                                text: '**Unconsciousness or unresponsiveness.** The person will be unresponsive to talking or stimulation.',
-                                subpoints: [
-                                    {
-                                        video: {
-                                            pageTitle: 'How To Simulate',
-                                            videoPath:
-                                                'https://youtu.be/uz5si4m44D8?si=ya4AqILDHKWJbnK60',
-                                            startTime: '00:00',
-                                            endTime: '0:59',
-                                        },
-                                    },
-                                    {
-                                        text: 'Try talking to the person, if they don’t respond perform a trapezius pinch on the person’s shoulder. This should wake someone up if they are sleeping. If they are unconscious, they will not respond.',
-                                        subpoints: [],
-                                    },
-                                ],
-                            },
-                            {
-                                text: '**Slowed or stopped breathing.** Normally, adults breathe around 12-20 times per minute, or about 1 time every 5 seconds.',
-                                subpoints: [
-                                    {
-                                        text: 'Get eye level with the person’s chest & place your ear near their nose and mouth. Listen for breath sounds while looking for chest rise and fall. Count their breaths.',
-                                        subpoints: [],
-                                    },
-                                ],
-                            },
-                        ],
-                    },
-                    {
-                        text: 'These are the main three signs that a person is overdosing. You can also check the person’s pulse.',
-                        subpoints: [
-                            {
-                                text: 'Only perform CPR if you are certified to do so and the person has no pulse.',
-                                subpoints: [
-                                    {
-                                        text: 'End Overdose does not train people on how to perform CPR.',
-                                        subpoints: [],
-                                    },
-                                ],
-                            },
-                        ],
-                    },
-                ],
-            },
-            {
-                text: 'The second step in responding to an opioid overdose is calling 911 and yelling for help. This is essential because 30-90 minutes after receiving naloxone, a person may return to a state of overdose.',
-                subpoints: [
-                    {
-                        text: '“SOMEONE IS UNRESPONSIVE AND NOT BREATHING”',
-                        subpoints: [],
-                    },
-                    {
-                        text: 'Don’t be afraid to call 911. Every state has a Good Samaritan Law, which protects individuals responding to emergencies. This law is different in every state, so make sure to consult your state’s law for more information.',
-                        subpoints: [],
-                    },
-                    {
-                        text: 'In California, the Good Samaritan Law is part of California Health & Safety Code § 1799.102 and states that ““No person who in good faith and not for compensation, renders emergency medical or non-medical care at the scene of an emergency shall be liable for any civil damages resulting from any act or omission.”',
-                        subpoints: [
-                            {
-                                text: 'This means that as long as it is not part of your job to respond to overdoses, you cannot be held civilly liable for anything you do or do not do when responding.',
-                                subpoints: [],
-                            },
-                            {
-                                text: 'Additionally, California’s AB 472 Controlled Substances: Overdose: Punishment states that individuals responding to overdoses are “...exempting drug possession or drug paraphernalia possession from criminal prosecution in situations involving medical emergencies.”',
-                                subpoints: [
-                                    {
-                                        text: 'This means you will not be punished for having drugs or drug paraphernalia as long as the amount you have is considered an amount for personal use.',
-                                        subpoints: [],
-                                    },
-                                ],
-                            },
-                        ],
-                    },
-                    {
-                        text: 'The Good Samaritan Law will not apply if…',
-                        subpoints: [
-                            {
-                                text: 'You could be charged with other crimes, like trespassing; you are on parole/probation',
-                                subpoints: [],
-                            },
-                            {
-                                text: 'You have more drugs than what is considered “possession for personal use”-- any amount suggesting trafficking or sales is still illegal',
-                                subpoints: [],
-                            },
-                            {
-                                text: 'You “obstruct medical or law enforcement personnel”-- once emergency responders arrive, let them take over responding.',
-                                subpoints: [],
-                            },
-                        ],
-                    },
-                ],
-            },
-            {
-                text: 'The third step in responding to an opioid overdose is administering naloxone to restore the person’s breathing.',
-                subpoints: [
-                    {
-                        video: {
-                            pageTitle: 'Learn to Rescue Breathe',
-                            videoPath:
-                                'https://youtu.be/QxpHP7bHzKc?si=eddrcVqL8PNaWb95',
-                            startTime: '00:00',
-                            endTime: '0:59',
-                        },
-                    },
-                    {
-                        text: 'First, lay the person on their back, then peel back the packages and place the naloxone spray in one nostril. Push the plunger all the way down to administer.',
-                        subpoints: [],
-                    },
-                    {
-                        text: 'Once you press the plunger, there will be nothing left in the dose.',
-                        subpoints: [],
-                    },
-                ],
-            },
-            {
-                text: 'If the person is not breathing, immediately after administering naloxone begin step four, rescue breathing.',
-                subpoints: [
-                    {
-                        text: 'To rescue breath, place the person flat on their back.',
-                        subpoints: [],
-                    },
-                    {
-                        text: 'Open their airway by tilting their chin back and form a seal by pinching their nose or using a breathing mask that cover their mouth and nose.',
-                        subpoints: [
-                            {
-                                text: 'Breathe into their mouth for one second and check that their chest rises and falls. Continue giving breaths every 5 seconds.',
-                                subpoints: [
-                                    {
-                                        text: 'If you do not see their chest rise and fall, adjust their airway by tilting their head or check for any obstructions.',
-                                        subpoints: [],
-                                    },
-                                ],
-                            },
-                        ],
-                    },
-                ],
-            },
-            {
-                text: 'Check often to see if the person has begun breathing on their own. If they are still not breathing 2-3 minutes after receiving naloxone, administer another dose in the opposite nostril.',
-                subpoints: [
-                    {
-                        text: 'Continue with rescue breathing as needed in between administering doses. Administer doses every 2-3 minutes until the person begins breathing on their own.',
-                        subpoints: [],
-                    },
-                ],
-            },
-            {
-                text: 'Step 5 will occur if the person begins to breathe on their own but remains unconscious.',
-                subpoints: [
-                    {
-                        text: 'Place the person into recovery position on their left side with their left arm and leg straight and their right arm and leg at a 90 degree angle.',
-                        subpoints: [],
-                    },
-                    {
-                        text: 'This position will ensure that the person will not choke if they need to vomit when waking up.',
-                        subpoints: [],
-                    },
-                ],
-            },
-        ],
-    },
-    // {
-    //     title: 'How To Stimulate',
-    //     content: [
-    //         {
-    //             video: {
-    //                 pageTitle: 'How To Stimulate',
-    //                 videoPath: 'https://www.youtube.com/watch?v=uz5si4m44D8',
-    //                 startTime: '00:00',
-    //                 endTime: '0:35',
-    //             },
-    //         },
-    //     ],
-    // },
-    // {
-    //     title: 'Learn to rescue breathe',
-    //     content: [
-    //         {
-    //             video: {
-    //                 pageTitle: 'Learn to rescue breathe',
-    //                 videoPath: 'https://www.youtube.com/watch?v=QxpHP7bHzKc',
-    //                 startTime: '00:00',
-    //                 endTime: '1:03',
-    //             },
-    //         },
-    //     ],
-    // },
-    {
-        title: 'Recap',
-        content: [
-            {
-                video: {
-                    pageTitle: 'Jack Shows How to Respond to an Overdose',
-                    videoPath:
-                        'https://youtu.be/JKwhqIHMD1E?si=TnOTk5xM-IQrQSr7',
-                    startTime: '00:00',
-                    endTime: '05:00',
-                },
-            },
-            {
-                text: 'Let’s go over how to respond 1 more time…',
-                subpoints: [
-                    {
-                        text: '1) Recognizing the symptoms',
-                        subpoints: [
-                            {
-                                text: 'Pinpoint pupils, slowed/stopped breathing, unconsciousness/unresponsiveness',
-                                subpoints: [],
-                            },
-                        ],
-                    },
-                    {
-                        text: '2) Calling 911 & yelling for help',
-                        subpoints: [
-                            {
-                                text: 'You are protected by the Good Samaritan Law!',
-                                subpoints: [],
-                            },
-                        ],
-                    },
-                    {
-                        text: '3) Administering naloxone',
-                        subpoints: [
-                            {
-                                text: 'Administer doses every 2-3 minutes until the person begins breathing.',
-                                subpoints: [],
-                            },
-                        ],
-                    },
-                    {
-                        text: '4) Performing rescue breathing',
-                        subpoints: [
-                            {
-                                text: 'Do this in between administering naloxone doses or if you don’t have any doses to administer.',
-                                subpoints: [],
-                            },
-                        ],
-                    },
-                    {
-                        text: '5) Placing the person in recovery position',
-                        subpoints: [
-                            {
-                                text: 'Do this if the person begins breathing but is still unconscious',
-                                subpoints: [],
-                            },
-                        ],
-                    },
-                ],
-            },
-            {
-                text: 'Here are some other things to consider…',
-                subpoints: [
-                    {
-                        text: 'Make sure the person overdosing gets to a hospital immediately. Naloxone wears off in 30-90 minutes. Try to stay with the person overdosing until emergency responders arrive.',
-                        subpoints: [],
-                    },
-                    {
-                        text: 'Naloxone has virtually NO effect on someone who is not overdosing. If you give it to someone accidentally, they will be fine, but make sure to keep your naloxone safe unless you are responding to an emergency,',
-                        subpoints: [],
-                    },
-                    {
-                        text: 'Naloxone can cause withdrawal symptoms which can make a person nauseous, irritable, and can cause cramps. It may also make the person want to take more drugs, but try not to let them.',
-                        subpoints: [],
-                    },
-                ],
-            },
-        ],
-    },
-];
-
-export default lessons;
